Split Actor into SignedObject and Actor.  Fixes to nomencalture in Messages.xml and relevant changes in source.  Deletion of MoveDirection and inclusion in MoveState as static constants.

+++ jgsf/src/main/java/com/jenjinstudios/jgsf/message/ExecutableStateChangeRequest.java
@@ -28,7 +28,7 @@ public class ExecutableStateChangeRequest extends WorldExecutableMessage
 	{
 		Actor player = getClientHandler().getActor();
 		MoveState newState;
-		int direction = (int) getMessage().getArgument("direction");
+		double direction = (double) getMessage().getArgument("direction");
 		double angle = (double) getMessage().getArgument("angle");
 		int stepsFromLast = (int) getMessage().getArgument("stepsUntilChange");
 		newState = new MoveState(direction, stepsFromLast, angle);
